E-Learning Platform/elearningplatform/requirements.txt
aioredis==1.3.1
amqp==5.2.0
appdirs==1.4.4
asgiref==3.7.2
astroid==2.4.2
async-timeout==4.0.3
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.2.0
black==21.7b0
celery==5.3.6
certifi==2020.12.5
cffi==1.14.4
channels==3.0.1
channels-redis==2.4.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.2
decorator==4.4.2
distlib==0.3.7
Django==3.1.3
django-bootstrap4==2.2.0
django-bootstrap5==23.4
django-braces==1.14.0
django-embed-video==1.4.0
django-memcache-status==2.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
factory-boy==3.3.0
Faker==20.0.0
filelock==3.12.4
flake8==3.9.2
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
isort==5.9.2
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.6.3
kombu==5.3.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
mypy-extensions==0.4.3
parso==0.7.1
pathspec==0.8.1
pbr==5.11.1
pickleshare==0.7.5
Pillow==8.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.6.1
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-memcached==1.59
pytz==2020.4
pywin32==228
PyYAML==6.0.1
pyzmq==19.0.2
redis==5.0.1
regex==2021.7.6
requests==2.23.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.5
sqlparse==0.4.1
stevedore==5.1.0
toml==0.10.1
tomli==1.0.4
tornado==6.0.4
traitlets==5.0.3
Twisted==21.2.0
twisted-iocpsupport==1.0.4
txaio==20.12.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.25.11
vine==5.1.0
virtualenv==20.24.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.13
wrapt==1.12.1
zope.interface==5.2.0
